+++ lib/server/hapi_plugins/api/index.js
@@ -3,19 +3,24 @@ var _ = require('lodash');
 var internals = require('./internals');
 var plugins = require('../../../helpers/plugin_api');
 
-exports.register = function (plugin, options, next) {
+exports.register = register;
+exports.register.attributes = {
+  name: 'api'
+};
+
+function register (server, options, next) {
   // allow clients to request a gzip response, even if the
   // Accept-Encoding headers is missing or mangled due to
   // faulty proxy servers
   // http://www.stevesouders.com/blog/2010/07/12/velocity-forcing-gzip-compression/
-  plugin.ext('onPreHandler', function maybeForceGzip (request, reply) {
+  server.ext('onPreHandler', function maybeForceGzip (request, reply) {
     if (request.query.force_gzip === 'true') {
       request.info.acceptEncoding = 'gzip';
     }
     reply.continue();
   });
 
-  plugin.route([
+  server.route([
     {
       method: ['GET', 'PUT', 'POST', 'DELETE', 'OPTIONS'],
       path: '/_api/{p*}',
@@ -101,8 +106,4 @@ exports.register = function (plugin, options, next) {
   ]);
 
   return next();
-};
-
-exports.register.attributes = {
-  pkg: require('./package.json')
-};
+}
