+++ lib/server/hapi_plugins/api/internals.js
