+++ lib/server/hapi_plugins/handle_404.js
@@ -1,11 +1,15 @@
 var fs = require('fs');
 var path = require('path');
 
-// https://github.com/spumko/hapi/blob/master/docs/Reference.md#plugin-interface
-exports.register = function (plugin, options, next) {
+exports.register = register;
+exports.register.attributes = {
+  name: 'handle_404'
+};
+
+function register (server, options, next) {
   var indexFile = path.join(options.app.www_root, 'index.html');
 
-  plugin.ext('onPostHandler', function (request, reply) {
+  server.ext('onPostHandler', function (request, reply) {
     var response = request.response;
 
     if (!response.isBoom) { return reply.continue(); }
@@ -21,10 +25,4 @@ exports.register = function (plugin, options, next) {
   });
 
   next();
-
-};
-
-exports.register.attributes = {
-  pkg: require('./package.json')
-};
-
+}
