+++ lib/server/hapi_plugins/logger.js
@@ -1,7 +1,12 @@
 var bytes = require('bytes');
 
-exports.register = function (plugin, options, next) {
-  plugin.on('response', function (request) {
+exports.register = register;
+exports.register.attributes = {
+  name: 'logger'
+};
+
+function register (server, options, next) {
+  server.on('response', function (request) {
     if (!options.app.verbose) {
       return;
     }
@@ -33,10 +38,5 @@ exports.register = function (plugin, options, next) {
     console.log(logStr);
   });
 
-  return next();
-};
-
-exports.register.attributes = {
-  pkg: require('./package.json')
-};
-
+  next();
+}
