--- lib/server/helpers/handle_404/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "handle_404",
-  "description": "",
-  "version": "1.0.0",
-  "main": "index",
-  "peerDependencies": {
-    "hapi": "2.x.x"
-  }
-}
-
