--- lib/server/helpers/logger/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "logger",
-  "description": "",
-  "version": "0.3.0",
-  "main": "index",
-  "peerDependencies": {
-    "hapi": "2.x.x"
-  }
-}
-
