+++ lib/server/index.js
@@ -4,8 +4,9 @@
 
 var Hapi = require('hapi');
 
-module.exports = function () {
+var hapi_plugins = require('./hapi_plugins');
 
+module.exports = function () {
   return function (env_config, callback) {
 
     var server = new Hapi.Server({
@@ -33,19 +34,11 @@ module.exports = function () {
       labels: ['admin']
     });
 
-    var hapi_plugins = [
-      './plugins/web',
-      './plugins/admin',
-      './plugins/api',
-      './helpers/logger',
-      './helpers/handle_404'
-    ];
-
     // register plugins against the server pack
     //
     hapi_plugins.forEach(function (plugin) {
       server.register({
-        register: require(plugin),
+        register: plugin,
         options: {
           app: env_config
         }
@@ -64,8 +57,5 @@ module.exports = function () {
 
       return callback();
     });
-
   };
-
 };
-
