+++ lib/server/nodejitsu_server.js
@@ -5,7 +5,6 @@
 var Hapi = require('hapi');
 
 module.exports = function (env_config, callback) {
-
   // skip this server if not on nodejitsu
   if (!env_config.run_nodejitsu_server) {
     return callback();
@@ -15,7 +14,7 @@ module.exports = function (env_config, callback) {
 
   server.connection({
     port: 8080,
-    labels: [ 'nodejitsu' ],
+    labels: ['nodejitsu'],
     routes: {
       cors: {
         override: false
@@ -51,6 +50,4 @@ module.exports = function (env_config, callback) {
   });
 
   server.start(callback);
-
 };
-
