--- lib/server/plugins/admin/package.json
@@ -1,9 +1,0 @@
-{
-  "name": "admin",
-  "description": "",
-  "version": "0.3.0",
-  "main": "index",
-  "peerDependencies": {
-    "hapi": "2.x.x"
-  }
-}
