--- lib/server/plugins/web/package.json
@@ -1,9 +1,0 @@
-{
-  "name": "web",
-  "description": "",
-  "version": "0.3.0",
-  "main": "index",
-  "peerDependencies": {
-    "hapi": "2.x.x"
-  }
-}
